+++ src/views/History/History.vue
@@ -6,7 +6,8 @@
       </ion-toolbar>
     </ion-header>
     <ion-content>
-      <ion-button @click="testCosign">承認テスト</ion-button>
+      <ion-button @click="reload">リロード</ion-button>
+      <br />
       <ion-label> 未承認トランザクション </ion-label>
       <ul id="example-1" style="list-style: none;">
         <li v-for="item in sample" :key="item.transactionInfo.hash">
@@ -53,24 +54,7 @@
     getTransactionList,
     getUnConfirmTransactionList,
   } from "@/util/symbol";
-  import { map, mergeMap } from "rxjs/operators";
-  import {
-    Account,
-    AggregateTransaction,
-    CosignatureSignedTransaction,
-    CosignatureTransaction,
-    NetworkType,
-    RepositoryFactoryHttp,
-    TransactionGroup,
-  } from "symbol-sdk";
-
-  const cosignAggregateBondedTransaction = (
-    transaction: AggregateTransaction,
-    account: Account
-  ): CosignatureSignedTransaction => {
-    const cosignatureTransaction = CosignatureTransaction.create(transaction);
-    return account.signCosignatureTransaction(cosignatureTransaction);
-  };
+  import { testCosign } from "@/util/cosign";
 
   export default {
     name: "Tab2",
@@ -105,6 +89,9 @@
       }
     },
     methods: {
+      reload() {
+        location.reload();
+      },
       async presentAlert(item: any) {
         const alert = await alertController.create({
           cssClass: "my-custom-class",
@@ -123,52 +110,12 @@
             {
               text: "OK",
               handler: () => {
-                this.testCosign(item.transactionInfo.hash);
+                testCosign(item.transactionInfo.hash);
               },
             },
           ],
         });
         return alert.present();
-      },
-      testCosign(hash: string) {
-        console.log("testCosign");
-        const networkType = NetworkType.TEST_NET;
-        // replace with private key
-        const privateKey =
-          "72D29CA347E87A7C4205D90BE51A800931D87402DF34A1FB5BD533BBC623E1A9";
-        const account = Account.createFromPrivateKey(privateKey, networkType);
-        // replace with node endpoint
-        // replace with transaction hash to cosign
-        const transactionHash = hash;
-        /* end block 02 */
-
-        /* start block 03 */
-        const nodeUrl = process.env.VUE_APP_WEB_SOCKET_URL;
-        const repositoryFactory = new RepositoryFactoryHttp(nodeUrl, {
-          websocketUrl: "ws://ngl-dual-601.testnet.symboldev.network:3000/ws",
-          websocketInjected: WebSocket,
-        });
-        const transactionHttp = repositoryFactory.createTransactionRepository();
-
-        transactionHttp
-          .getTransaction(transactionHash, TransactionGroup.Partial)
-          .pipe(
-            map((transaction) =>
-              cosignAggregateBondedTransaction(
-                transaction as AggregateTransaction,
-                account
-              )
-            ),
-            mergeMap((cosignatureSignedTransaction) =>
-              transactionHttp.announceAggregateBondedCosignature(
-                cosignatureSignedTransaction
-              )
-            )
-          )
-          .subscribe(
-            (announcedTransaction) => console.log(announcedTransaction),
-            (err) => console.error(err)
-          );
       },
     },
   };
